Conflicts:
	source/scene/gamescene.cpp

+++ source/controller/cameracontroller.cpp
@@ -36,20 +36,20 @@ void CameraController::LookAt(const Vector3f &pos)
 	s32 y = static_cast<s32>(p.getY());
 	f32 z = p.getZ();
 
-	s32 halfScreenWidth = kHalfViewW;//pScreen->GetWidth() / 2;
-	s32 halfScreenHeight = kHalfViewH;//pScreen->GetHeight() / 2;
+	s32 halfScreenWidth = s32(kHalfViewW);//pScreen->GetWidth() / 2;
+	s32 halfScreenHeight = s32(kHalfViewH);//pScreen->GetHeight() / 2;
 
-	if (x < cArea.x1 - 16)
-		x = cArea.x1 - 16;
+	if (x < s32(cArea.x1 - 16))
+		x = s32(cArea.x1 - 16);
 
-	if (y < cArea.y1 - 16)
-		y = cArea.y1 - 16;
+	if (y < s32(cArea.y1 - 16))
+		y = s32(cArea.y1 - 16);
 
-	if (y > cArea.y2 - 16 - halfScreenHeight * 2)
-		y = cArea.y2 - 16 - halfScreenHeight * 2;
+	if (y > s32(cArea.y2 - 16 - halfScreenHeight * 2))
+		y = s32(cArea.y2 - 16 - halfScreenHeight * 2);
 
-	if (x > (cArea.x2 - 16 - halfScreenWidth * 2))
-		x = cArea.x2 - 16 - halfScreenWidth * 2;
+	if (x > s32(cArea.x2 - 16 - halfScreenWidth * 2))
+		x = s32(cArea.x2 - 16 - halfScreenWidth * 2);
 
 	Vector3f np(static_cast<f32>(x), static_cast<f32>(y), z);
 	pCamera->SetPosition(np);
